requirements.txt
# No external packages required for this project
# All dependencies are part of Python's standard library:
# - tkinter (GUI framework)
# - sqlite3 (database)
# - csv (for export functionality)
# - datetime (for date handling)
# - re (for email validation)
# - os (for file operations)
Flask
flask-cors
pyjwt
werkzeug